refactor: move lastSeenCustomRange into SystemsViewFetchParams
