+++ src/components/SystemsView/types.ts
@@ -2,6 +2,11 @@ import type { ISortBy } from '@patternfly/react-table';
 
 export type SortDirection = ISortBy['direction'];
 
+export type LastSeenCustomRange = {
+  start?: string;
+  end?: string;
+} | null;
+
 export type SystemsViewFetchParams<
   TFilters extends Record<string, unknown> = Record<string, unknown>,
 > = {
@@ -10,4 +15,5 @@ export type SystemsViewFetchParams<
   sortBy: string | undefined;
   direction: SortDirection | undefined;
   filters: TFilters;
+  lastSeenCustomRange: LastSeenCustomRange;
 };
